Cmake looks for 'invaders.rom' in the root of the project directory (same dir as CMakeLists.txt) and copies it into the build directory.
When a wrapper is created, it then loads this rom into the RAM block it has created.

The reason for this is that fopen can only open files in the cwd, or else we have to use a full path. Using a full path is dependent on the user's directory structure, so I just have Cmake move the file to the right location at build time.