Integrate GUI code with Emulator Code (#8)
Merging the GUI code with emulator code. Emulator code now runs in emulatorWrapper object.
Current status - runs through instructions up until the loop where emulator is expecting a hardware interrupt from video system.

* Copy over GUI repo files
Add a gitignore for Qt files
Deleted main.c and makefile since we will be using main.cpp for the app and the makefile will be generated by cmake going forward

* Forgot to commit inputManager and keymap

* Forgot to commit main.cpp

* Copy over GUI repo files
Add a gitignore for Qt files
Deleted main.c and makefile since we will be using main.cpp for the app and the makefile will be generated by cmake going forward

* Update CMakeLists.txt to include c-code emulator related files from emulator repo, move all to project sources

* Update gitignore to include build directory

* Update disassembler.c

fixed indentations

* Update disassembler.h

added include C++

* Update emulator.h

* Update mem_utils.h

* Update memory.h

* Added emulator.h to sources in CMakeLists.txt

* Important! - Must add C to list of project languages so compiler will handle C files correctly

* Emulator wrapper now has a mem_block to load the ROM into

* Change memory block name to 'ram' for clarity

* Emulator wrapper now has a mem_block member called 'ram' to load the ROM into

* Changed the dummy ioport reader so it doesn't spam qDebug with empty strings when no ports are active

* Update .gitignore to include rom files

* Update .gitignore to include rom files

* Cmake looks for 'invaders.rom' in the root of the project directory (same dir as CMakeLists.txt) and copies it into the build directory.

When a wrapper is created, it then loads this rom into the RAM block it has created.

The reason for this is that fopen can only open files in the cwd, or else we have to use a full path. Using a full path is dependent on the user's directory structure, so I just have Cmake move the file to the right location at build time.

* Cmake looks for 'invaders.rom' in the root of the project directory (same dir as CMakeLists.txt) and copies it into the build directory.

When a wrapper is created, it then loads this rom into the RAM block it has created.

The reason for this is that fopen can only open files in the cwd, or else we have to use a full path. Using a full path is dependent on the user's directory structure, so I just have Cmake move the file to the right location at build time.

* Set up 'ram' and 'state' members for emulatorWrapper class.

Successfully loads ROM contents into ram:
NOTE: in order for Cmake to place it in the proper location at build time, the invaders.rom file should be placed in the root directory of the project

Currently just runs a dummy emulator loop that displays what gameplay keys are being pressed.

Trying to tie together state.memory and ram.mem, but it crashes everything. See TODOs in emulatorWrapper.cpp.

* Fixed pointer assignment issue
TODO: emulate8080cpu still crashes things

* Fixed pointer assignment issue
TODO: emulate8080cpu still crashes things

* Fixed error - no definition for 'state' in emulatorWrapper.h

* Reduced delay timer between cycles so the program has time to output
before crashing. Removed dummy keystroke reporter to declutter output.

Status - running until crash. Should be able to debug emulate8080
function now.

* Cleaned up unneeded debug calls and documentation

* Initialize stack and program counters to avoid possible issues

* Implemented missing mov instruction

* Fixed crash when attempting to disassemble and display MOV instructions
- the definitions were inside the case code and never executed.

* changed MOV case ordering and comments

* Removed timing code from emulatorWrapper until we find a cross-platform
solution

* Removed resolved TODO from disassemble - it handles MOV cases fine now

---------

Co-authored-by: Noah Freeman <114368114+librepax@users.noreply.github.com>
Co-authored-by: Noah Freeman <freemnoa@oregonstate.edu>